requirements.txt
fastapi
uvicorn
openai
streamlit
chromadb
sentence-transformers
python-dotenv
requests
google-generativeai
